+++ packages/ide-extension/test/types/index.test.ts
@@ -1,0 +1,5 @@
+import * as types from '../../src/types';
+
+test('Check types are defined', () => {
+    expect(types).toBeDefined();
+});