+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap/guided-answers-extension-core
 
+## 1.4.1
+
+### Patch Changes
+
+-   0d5eb00: Setup telemetry for product improvement
+-   Updated dependencies [0d5eb00]
+    -   @sap/guided-answers-extension-types@1.3.1
+
 ## 1.4.0
 
 ### Minor Changes